+++ server/src/main/java/org/opensearch/action/admin/cluster/node/stats/NodesStatsRequest.java
@@ -38,6 +38,7 @@ import org.opensearch.common.io.stream.StreamInput;
 import org.opensearch.common.io.stream.StreamOutput;
 import java.io.IOException;
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.SortedSet;
@@ -53,6 +54,7 @@ public class NodesStatsRequest extends BaseNodesRequest<NodesStatsRequest> {
 
     private CommonStatsFlags indices = new CommonStatsFlags();
     private final Set<String> requestedMetrics = new HashSet<>();
+    private final Set<String> restActionsFilters = new HashSet<>();
 
     public NodesStatsRequest() {
         super((String[]) null);
@@ -64,6 +66,7 @@ public class NodesStatsRequest extends BaseNodesRequest<NodesStatsRequest> {
         indices = new CommonStatsFlags(in);
         requestedMetrics.clear();
         requestedMetrics.addAll(in.readStringList());
+        restActionsFilters.addAll(in.readStringList());
     }
 
     /**
@@ -80,6 +83,7 @@ public class NodesStatsRequest extends BaseNodesRequest<NodesStatsRequest> {
     public NodesStatsRequest all() {
         this.indices.all();
         this.requestedMetrics.addAll(Metric.allMetrics());
+        this.clearRestActionsFilters();
         return this;
     }
 
@@ -89,6 +93,7 @@ public class NodesStatsRequest extends BaseNodesRequest<NodesStatsRequest> {
     public NodesStatsRequest clear() {
         this.indices.clear();
         this.requestedMetrics.clear();
+        this.clearRestActionsFilters();
         return this;
     }
 
@@ -180,11 +185,26 @@ public class NodesStatsRequest extends BaseNodesRequest<NodesStatsRequest> {
         }
     }
 
+    public void addRestActionsFilters(String... filters) {
+        if (null == filters || filters.length < 1) return;
+        Arrays.stream(filters).forEach(restActionsFilters::add);
+    }
+
+    public void clearRestActionsFilters() {
+        if (null == restActionsFilters || restActionsFilters.isEmpty()) return;
+        restActionsFilters.clear();
+    }
+
+    public Set<String> getRestActionsFilters() {
+        return Collections.unmodifiableSet(restActionsFilters);
+    }
+
     @Override
     public void writeTo(StreamOutput out) throws IOException {
         super.writeTo(out);
         indices.writeTo(out);
         out.writeStringArray(requestedMetrics.toArray(new String[0]));
+        out.writeStringArray(restActionsFilters.toArray(new String[0]));
     }
 
     /**
@@ -207,7 +227,8 @@ public class NodesStatsRequest extends BaseNodesRequest<NodesStatsRequest> {
         SCRIPT_CACHE("script_cache"),
         INDEXING_PRESSURE("indexing_pressure"),
         SHARD_INDEXING_PRESSURE("shard_indexing_pressure"),
-        SEARCH_BACKPRESSURE("search_backpressure");
+        SEARCH_BACKPRESSURE("search_backpressure"),
+        REST_ACTIONS("rest_actions");
 
         private String metricName;