+++ server/src/test/java/org/opensearch/rest/action/admin/cluster/RestNodesStatsActionTests.java
@@ -40,6 +40,7 @@ import org.opensearch.test.rest.FakeRestRequest;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.Map;
 import java.util.Set;
 
 import static org.hamcrest.CoreMatchers.containsString;
@@ -166,4 +167,26 @@ public class RestNodesStatsActionTests extends OpenSearchTestCase {
         );
     }
 
+    public void testRestActionsMetricWithFilterOnAllRequest() throws IOException {
+        final RestRequest request = new FakeRestRequest.Builder(xContentRegistry()).withPath("/_nodes/stats")
+            .withParams(Map.of("metric", "_all", "rest_actions_filters", "action_a, action_b"))
+            .build();
+
+        expectThrows(IllegalArgumentException.class, () -> action.prepareRequest(request, mock(NodeClient.class)));
+    }
+
+    public void testRestActionsFilterWithoutRestActionsMetricRequest() throws IOException {
+        final RestRequest request = new FakeRestRequest.Builder(xContentRegistry()).withPath("/_nodes/stats")
+            .withParams(
+                Map.of(
+                    "metric",
+                    randomSubsetOf(1, RestNodesStatsAction.METRICS.keySet()).get(0),
+                    "rest_actions_filters",
+                    "action_a, action_b"
+                )
+            )
+            .build();
+
+        expectThrows(IllegalArgumentException.class, () -> action.prepareRequest(request, mock(NodeClient.class)));
+    }
 }