+++ test/framework/src/main/java/org/opensearch/test/InternalTestCluster.java
@@ -2686,7 +2686,9 @@ public final class InternalTestCluster extends TestCluster {
                     false,
                     false,
                     false,
-                    false
+                    false,
+                    false,
+                    null
                 );
                 assertThat(
                     "Fielddata size must be 0 on node: " + stats.getNode(),